+++ src/main/app/util/PseudoMaps/SettingsWrapper.ts
@@ -22,7 +22,8 @@ export class SettingsWrapper extends TypedMap<Settings> implements Persistable {
   public static readonly DEFAULT: Settings = {
     debugMode: { enabled: false, anyDomainOverrideUrl: "U131364=https://templateassisted.myassn.com/", versionOverride: "1.1.0" },
     updateCheck: { enabled: true, showNotifications: true },
-    squelch: { pullComplete: false, pushComplete: false }
+    squelch: { pullComplete: false, pushComplete: false },
+    autoSave: { enabled: false }
   };
 
   constructor() {