+++ src/main/resources/constants/SettingsKeys.ts
@@ -8,4 +8,6 @@ export namespace SettingsKeys {
   export const DEBUG_MODE = "debugMode";
   export const UPDATE_CHECK = "updateCheck";
   export const UPDATE_CHECK_SHOW_NOTIFICATIONS = "updateCheck.showNotifications";
+  export const AUTO_SAVE = "autoSave";
+  export const AUTO_SAVE_ENABLED = "autoSave.enabled";
 }